+++ src/LaunchpadServiceProvider.php
@@ -36,6 +36,7 @@ class LaunchpadServiceProvider extends PackageServiceProvider
                 '2024_01_01_000001_create_launchpad_tables',
                 '2024_01_01_000002_seed_default_launchpad_home',
                 '2024_01_01_000003_add_panel_id_to_launchpad_spaces_table',
+                '2024_01_01_000004_make_launchpad_user_ids_string',
             ])
             ->runsMigrations();
     }